EES-3777 Refactor `ResultBuilder` to `ObservationViewModelBuilder`
